feat(availableBalance): improved logic and fixed comments FEE-723
