+++ packages/react/src/domain/model/index.ts
@@ -1,5 +1,3 @@
 export * from './Country';
 export * from './Address';
 export * from './TokenConversion';
-export * from './TokenSwapQuote';
-export * from './AvailableBalance';
