--- packages/react/src/providers/helio/actions/index.ts
@@ -1,1 +1,0 @@
-export * from './fetchAvailableBalances';
