+++ packages/react/src/ui-kits/index.tsx
@@ -3,7 +3,7 @@ export type { PriceBannerProps } from './priceBanner';
 export type { ElementContainerProps } from './elementContainer';
 export type { Option as SelectBoxOption } from './selectbox';
 
-export { default as Button } from './button';
+export { Button } from './button';
 export { default as Input } from './input';
 export { default as PhoneNumberInput } from './phoneNumberInput';
 export { default as NumberInput } from './numberInput';
@@ -17,4 +17,4 @@ export { default as ConnectButton } from './connectButton';
 export { default as ExplorerLink } from './explorerLink';
 
 export * from './tooltip';
-export * from './buttonWithTooltip'; // eslint-disable-line import/no-cycle
+export * from './buttonWithTooltip';
