+++ packages/sdk/src/infrastructure/solana-utils/index.ts
@@ -1,4 +1,3 @@
 export * from './payment/paylink/PaylinkSubmitService';
 export * from './payment/paystream/PaystreamStartService';
 export * from './payment/paystream/PaystreamCancelService';
-export * from './GetTokenBalances';
